python/model/requirements.txt
boto3
pydantic
tensorflow-cpu